+++ duckscript_sdk/src/test/mod.rs
@@ -139,7 +139,7 @@ impl Command for OnErrorCommand {
 
 pub(crate) enum CommandValidation {
     None,
-    Ignore,
+    PositiveNumber(String),
     Match(String, String),
     Contains(String, String),
     Any(String, Vec<String>),
@@ -237,7 +237,13 @@ pub(crate) fn run_script_and_validate(
                         &values
                     )
                 }
-                CommandValidation::Ignore => (),
+                CommandValidation::PositiveNumber(key) => {
+                    assert!(!context.variables.is_empty());
+
+                    let var_value = context.variables.get(&key).unwrap();
+                    let numeric_value: u128 = var_value.parse().unwrap();
+                    assert!(numeric_value > 0)
+                }
             };
 
             context